+++ src/CrossWindowProvider/tests/Cancel.test.ts
@@ -1,5 +1,4 @@
 import { getWalletWindowMock, WalletWindowMockType } from '../../test-utils';
-import { WindowManager } from '../../WindowManager/WindowManager';
 import { CrossWindowProvider } from '../CrossWindowProvider';
 
 describe('CrossWindowProvider Login', () => {
@@ -9,11 +8,11 @@ describe('CrossWindowProvider Login', () => {
 
   beforeEach(() => {
     walletWindowMock = getWalletWindowMock();
-    WindowManager.getInstance().postMessage = jest
+    CrossWindowProvider.getInstance().getWindowManager().postMessage = jest
       .fn()
       .mockImplementation(() => undefined);
 
-    WindowManager.getInstance().isWalletOpened = jest
+    CrossWindowProvider.getInstance().getWindowManager().isWalletOpened = jest
       .fn()
       .mockImplementation(() => true);