+++ src/CrossWindowProvider/tests/SignMessage.test.ts
@@ -1,6 +1,5 @@
 import { SignableMessage } from '@multiversx/sdk-core/out';
 import { getWalletWindowMock, WalletWindowMockType } from '../../test-utils';
-import { WindowManager } from '../../WindowManager/WindowManager';
 import { CrossWindowProvider } from '../CrossWindowProvider';
 
 describe('CrossWindowProvider Login', () => {
@@ -10,7 +9,7 @@ describe('CrossWindowProvider Login', () => {
 
   beforeEach(() => {
     walletWindowMock = getWalletWindowMock();
-    WindowManager.getInstance().postMessage = jest
+    CrossWindowProvider.getInstance().getWindowManager().postMessage = jest
       .fn()
       .mockImplementation(() => ({
         payload: {