+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Implement IFrameProvider](https://github.com/multiversx/mx-wallet-dapp/pull/46)
+
 ## [[0.1.3](https://github.com/multiversx/mx-wallet-dapp/pull/45)] - 2024-05-29
 - [Changed lit library to vanilla](https://github.com/multiversx/mx-wallet-dapp/pull/44)